+++ Samples~/Playtesting/PlaytestingCanvas.cs
@@ -31,12 +31,14 @@ namespace Agava.YandexGames.Samples
 
         private IEnumerator Start()
         {
-#if !UNITY_WEBGL || UNITY_EDITOR
-            yield break;
-#endif
+            if (!YandexGamesSdk.IsRunningOnYandex)
+                yield break;
 
+            CanvasGroup canvasGroup = GetComponent<CanvasGroup>();
+            canvasGroup.interactable = false;
             // Always wait for it if invoking something immediately in the first scene.
             yield return YandexGamesSdk.Initialize();
+            canvasGroup.interactable = true;
 
             if (PlayerAccount.IsAuthorized == false)
                 PlayerAccount.StartAuthorizationPolling(1500);